rockblock-host/requirements.txt
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mysql-connector==2.1.4
pbr==3.1.1
pip==9.0.1
PyMySQL==0.7.11
setuptools==36.2.6
six==1.10.0
SQLAlchemy==1.1.13
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
Werkzeug==0.12.2
wheel==0.29.0
